+++ macos-leo/app/Components/Types/MusicType.ts
@@ -1,0 +1,7 @@
+export interface MusicItem {
+  id: number;
+  title: string;
+  writter: string;
+  imageSrc: string;
+  music: string;
+}